ejercicio area -PR2 -CONTROL-COMMIT

+++ Exercises/MODULE-2 ALGORITHMS/PR2/quilometres.cpp
@@ -14,7 +14,7 @@ int main(void) {
     cout << "Dame los kilometros: "; 
     cin >> kilometros;
 
-    int millas = kilometros * MILLA;
+    float millas = kilometros * MILLA;
 
     cout << "Tus kilometros en millas son: " << millas << " millas\n";
